Refactored e2e tests for clarity
